Enhancements

+++ twt_img.py
@@ -15,10 +15,11 @@ class Downloader:
 
     def __init__(self, api_key, api_secret):
         self.bearer_token = self.bearer(api_key, api_secret)
+        #print ('Bearer token is ' + self.bearer_token)
         self.last_tweet = None
         self.count = 0
 
-    def download_images(self, user, save_dest, size='large', rts=False):
+    def download_images(self, user, save_dest, size='large', limit=3200, rts=False):
         '''Download and save images that user uploaded.
 
         Args:
@@ -31,8 +32,12 @@ class Downloader:
         if not os.path.isdir(save_dest):
             raise InvalidDownloadPathError()
 
-        tweets = self.get_tweets(user, self.last_tweet, rts)
-        while len(tweets) > 0:
+        num_tweets_checked = 0
+        tweets = self.get_tweets(user, self.last_tweet, limit, rts)
+        if not tweets:
+            print ("Got an empty list of tweets")
+
+        while len(tweets) > 0 and num_tweets_checked < limit:
             for tweet in tweets:
                 # create a file name using the timestamp of the image
                 timestamp = dateutil.parser.parse(tweet['created_at']).timestamp()
@@ -41,8 +46,10 @@ class Downloader:
                 # save the image
                 image = self.extract_image(tweet)
                 self.save_image(image, save_dest, fname, size)
+                num_tweets_checked += 1
                 self.last_tweet = tweet['id']
-            tweets = self.get_tweets(user, self.last_tweet)
+
+            tweets = self.get_tweets(user, self.last_tweet, count=limit)
 
     def bearer(self, key, secret):
         '''Receive the bearer token and return it.
@@ -70,7 +77,7 @@ class Downloader:
         else:
             raise BearerTokenNotFetchedError()
 
-    def get_tweets(self, user, start=None, rts=False):
+    def get_tweets(self, user, start=None, count=200, rts=False):
         '''Download user's tweets and return them as a list.
 
         Args:
@@ -85,7 +92,7 @@ class Downloader:
         headers = {
             'Authorization': 'Bearer {}'.format(bearer_token)
         }
-        payload = {'screen_name': user, 'count': 200, 'include_rts': rts}
+        payload = {'screen_name': user, 'count': count, 'include_rts': rts}
         if start:
             payload['max_id'] = start
 
@@ -101,6 +108,7 @@ class Downloader:
                 print('Got ' + str(len(tweets)) + ' tweets')
                 return tweets if not start else tweets[1:]
         else:
+            print ('An error occurred with the request, status code was ' + str(r.status_code))
             return []
 
     def extract_image(self, tweet):
@@ -127,19 +135,24 @@ class Downloader:
         '''
 
         if image:
-            print('Saving ' + image)
 
             # image's path with a new name
             ext = os.path.splitext(image)[1]
             save_dest = os.path.join(path, timestamp + ext)
 
-            # save the image in the specified directory
-            r = requests.get(image + ':' + size, stream=True)
-            if r.status_code == 200:
-                with open(save_dest, 'wb') as f:
-                    r.raw.decode_content = True
-                    shutil.copyfileobj(r.raw, f)
-                self.count += 1
+            # save the image in the specified directory (or don't)
+            if not (os.path.exists(save_dest)):
+                print('Saving ' + image)
+
+                r = requests.get(image + ':' + size, stream=True)
+                if r.status_code == 200:
+                    with open(save_dest, 'wb') as f:
+                        r.raw.decode_content = True
+                        shutil.copyfileobj(r.raw, f)
+                    self.count += 1
+
+            else:
+                print('Skipping ' + image + ' because it was already dowloaded')
 
 
 if __name__ == '__main__':
@@ -147,7 +160,8 @@ if __name__ == '__main__':
     parser.add_argument('user_id', help='an ID of a twitter user')
     parser.add_argument('dest', help='specify where to put images')
     parser.add_argument('-c', '--confidentials', help='a json file containing a key and a secret')
-    parser.add_argument('-s', '--size',  help='specify the size of images', default='large', choices=['large', 'medium', 'small', 'thumb'])
+    parser.add_argument('-s', '--size',  help='specify the size of images', default='large', choices=['large', 'medium', 'small', 'thumb', 'orig'])
+    parser.add_argument('-l', '--limit', type=int, help='the maximum number of tweets to check (most recent first)', default=3200)
     parser.add_argument('--rts', help='save images contained in retweets', action="store_true")
     args = parser.parse_args()
 
@@ -162,4 +176,4 @@ if __name__ == '__main__':
         raise ConfidentialsNotSuppliedError()
 
     downloader = Downloader(api_key, api_secret)
-    downloader.download_images(args.user_id, args.dest, args.size, args.rts)
+    downloader.download_images(args.user_id, args.dest, args.size, args.limit, args.rts)
